# Timing benchmark tweaks and fixes