This is a bunch of independent tweaks to the `test/benchmarks/` runner script to make using it for evaluating optimizer sequences a little less tedious. Each could really be a separate small  PR.
- The script would not continue when legacy compilation failed with "Stack too deep".
- Bytecode size is a useful additional data point that we can easily get here.
- Putting the values manually in a table gets old quickly. Now the script prints a table I can just copy paste into a comment on Github.
- The lighter tests now run first so you can start seeing results immediately. Also, now the script does not wait with printing legacy results until IR finishes.
- solc binary is now a parameter so that I can easily use system-wide solc. For testing sequences I find it easier to use the release binary and just give it different `--yul-optimizations`.