@cameel 2024-04-12
### Before
```
=======================================================
            chains.sol
-------------------------------------------------------
legacy pipeline took 0.96 seconds to execute.
via-ir pipeline took 69.23 seconds to execute.
=======================================================
=======================================================
            OptimizorClub.sol
-------------------------------------------------------
legacy pipeline took Command exited with non-zero status 1
2.78 seconds to execute.
via-ir pipeline took 16.33 seconds to execute.
=======================================================
=======================================================
            verifier.sol
-------------------------------------------------------
legacy pipeline took 0.74 seconds to execute.
via-ir pipeline took 3.21 seconds to execute.
=======================================================

=======================================================
Warnings and errors generated during run:
=======================================================
```

### After
```
| File                 | Pipeline | Bytecode size | Time     | Exit code |
|----------------------|----------|--------------:|---------:|----------:|
| `verifier.sol`       | legacy   |    4983 bytes |   0.79 s |         0 |
| `verifier.sol`       | via-ir   |    4408 bytes |   3.51 s |         0 |
| `OptimizorClub.sol`  | legacy   |       0 bytes |   2.97 s |         1 |
| `OptimizorClub.sol`  | via-ir   |   22427 bytes |  16.99 s |         0 |
| `chains.sol`         | legacy   |    5911 bytes |   0.88 s |         0 |
| `chains.sol`         | via-ir   |   21443 bytes |  66.46 s |         0 |

=======================================================
Warnings and errors generated during run:
=======================================================
```

### Formatted
| File                 | Pipeline | Bytecode size | Time     | Exit code |
|----------------------|----------|--------------:|---------:|----------:|
| `verifier.sol`       | legacy   |    4983 bytes |   0.79 s |         0 |
| `verifier.sol`       | via-ir   |    4408 bytes |   3.51 s |         0 |
| `OptimizorClub.sol`  | legacy   |       0 bytes |   2.97 s |         1 |
| `OptimizorClub.sol`  | via-ir   |   22427 bytes |  16.99 s |         0 |
| `chains.sol`         | legacy   |    5911 bytes |   0.88 s |         0 |
| `chains.sol`         | via-ir   |   21443 bytes |  66.46 s |         0 |